refactor(api): extract scanner handler functions into a module
